+++ .buildkite/commands/build.sh
@@ -1,4 +1,4 @@
-#!/bin/bash -u
+#!/bin/bash -eu
 
 if "$(dirname "${BASH_SOURCE[0]}")/should-skip-job.sh" --job-type build; then
   exit 0