+++ podcasts/End of Year/Stories/2024/ListeningTime2024Story.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import EndOfYear
 
 struct ListeningTime2024Story: ShareableStory {
     @Environment(\.renderForSharing) var renderForSharing: Bool