+++ podcasts/End of Year/Stories/2025/CompletionRate2025Story.swift
@@ -2,6 +2,7 @@ import SwiftUI
 import PocketCastsServer
 import PocketCastsDataModel
 import Lottie
+import EndOfYear
 
 struct CompletionRate2025Story: ShareableStory {
     @Environment(\.renderForSharing) var renderForSharing: Bool