+++ podcasts/End of Year/Stories/2025/EpilogueStory2025.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import EndOfYear
 
 struct EpilogueStory2025: StoryView {
     private let foregroundColor = Color.white