+++ podcasts/End of Year/Stories/2025/NumberListened2025.swift
@@ -2,6 +2,7 @@ import SwiftUI
 import PocketCastsDataModel
 import Lottie
 import Combine
+import EndOfYear
 
 class StepCounter: ObservableObject {