+++ podcasts/End of Year/StoriesModel.swift
@@ -1,6 +1,7 @@
 import Combine
 import PocketCastsServer
 import SwiftUI
+import EndOfYear
 
 @MainActor
 class StoriesModel: ObservableObject {